Move SimpleEditPanel to its own file
